Refactor JdbcTimestampBoundStoreTest
