+++ atlasdb-dbkvs-tests/src/test/java/com/palantir/atlasdb/keyvalue/dbkvs/PostgresDbTimestampBoundStoreTest.java
@@ -15,11 +15,20 @@
  */
 package com.palantir.atlasdb.keyvalue.dbkvs;
 
+import org.junit.After;
+
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionManagerAwareDbKvs;
 import com.palantir.atlasdb.keyvalue.dbkvs.timestamp.PostgresDbTimestampBoundStore;
+import com.palantir.atlasdb.timestamp.AbstractDbTimestampBoundStoreTest;
 import com.palantir.timestamp.TimestampBoundStore;
 
 public class PostgresDbTimestampBoundStoreTest extends AbstractDbTimestampBoundStoreTest {
+    ConnectionManagerAwareDbKvs kvs;
+
+    @After
+    public void tearDown() throws Exception {
+        kvs.close();
+    }
 
     @Override
     protected TimestampBoundStore getTimestampBoundStore() {
